src/question_bank/requirements.txt
google-cloud-storage
vertexai
pandas
pytest
pytest-cov